requirements.txt
streamlit==1.33.0
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-text-splitters==0.2.2
groq==0.5.0
faiss-cpu==1.13.2
pypdf
python-dotenv
sentence-transformers
torch==2.2.2
numpy==1.26.4
httpx==0.27.0